feat: basic templating

+++ src/commands/build.rs
@@ -1,28 +1,72 @@
 use std::error::Error;
 use std::fs;
 use std::fs::{copy, create_dir_all, DirEntry, try_exists};
-use std::io::ErrorKind::NotFound;
-use std::path::{Path, PathBuf};
-use pulldown_cmark::{html, Options, Parser};
+use std::path::PathBuf;
+use pathdiff::diff_paths;
+use tera::{Context, Tera};
+use crate::structures::Post;
 
 const POST_DIR: &str = "posts";
 const SITE_DIR: &str = "site";
+const THEME_DIR: &str = "themes";
+const THEME: &str = "aoike";
 
 /// Perform a full site build
 pub fn build(src_dir: &PathBuf) {
     let post_dir = src_dir.join(POST_DIR);
     let site_dir = src_dir.join(SITE_DIR);
-    println!("building with src_dir={src_dir:?}");
+    let theme_dir = src_dir.join(THEME_DIR).join(THEME);
+    println!("building with src_dir={:?}", src_dir);
 
-    fs::remove_dir_all(&site_dir).expect("Clean failed");
+    // Clean up
+    if try_exists(&site_dir).unwrap_or(false) {
+        fs::remove_dir_all(&site_dir).expect("Clean failed");
+    }
 
-    let files = get_files(&post_dir).expect("Get files failed");
-    println!("{files:?}");
+    let template_path = theme_dir.join("**/*.html");
+    println!("Loading templates from {:?}", template_path);
+    let mut tera = match Tera::new(template_path.to_str().unwrap()) {
+        Ok(t) => t,
+        Err(e) => {
+            panic!("Parsing error(s): {}", e);
+        }
+    };
+    tera.autoescape_on(vec![".html"]);
+
+    // Copy theme files
+    let theme_files = get_files(&theme_dir).expect("Get theme files failed");
+    for entry in theme_files {
+        let src_path = entry.path();
+        let dst_path = site_dir.join(src_path.strip_prefix(&theme_dir).unwrap());
+
+        let parent = dst_path.parent().expect("Get Parent failed");
+        if !try_exists(parent).expect("Try exist failed") {
+            create_dir_all(parent).expect("Create dir failed");
+        }
+
+        if entry.path().extension().map(|s| s == "html").unwrap_or(false) {
+            continue;
+        }
+        copy(src_path, dst_path).expect("Copy theme file failed");
+    }
+
+    let mut posts: Vec<Post> = Vec::new();
+
+    // Build posts
+    let files = get_files(&post_dir).expect("Get posts files failed");
     for entry in files {
         let src_path = entry.path();
-        // println!("{path:?}");
+        let rel_root_path = diff_paths(&post_dir, &src_path.parent().unwrap()).expect("Calc relative root path failed");
+        let rel_root_path = rel_root_path.to_str().map(|s| {
+            if s.len() == 0 {
+                "."
+            } else {
+                s
+            }
+        }).unwrap();
+
+        println!("src_path: {:?}, post_dir: {:?}, rel_root_path: {:?}", src_path, post_dir, rel_root_path);
         let dst_path = site_dir.join(src_path.strip_prefix(&post_dir).unwrap());
-        // println!("{path:?}");
 
         let parent = dst_path.parent().expect("Get Parent failed");
         if !try_exists(parent).expect("Try exist failed") {
@@ -30,19 +74,17 @@ pub fn build(src_dir: &PathBuf) {
         }
 
         if entry.path().extension().map(|s| s == "md").unwrap_or(false) {
+            let post = Post::from_entry(entry);
             let dst_path = dst_path.with_extension("html");
-            let markdown = fs::read_to_string(src_path).expect("Failed to read markdown file");
-            let mut options = Options::empty();
-            options.insert(Options::ENABLE_STRIKETHROUGH);
-            options.insert(Options::ENABLE_TABLES);
-            options.insert(Options::ENABLE_TASKLISTS);
-            let parser = Parser::new_ext(&markdown, options);
+            println!("Building Post: {:?} to {:?}", post.file_path(), dst_path);
 
-            let mut html = String::new();
-            html::push_html(&mut html, parser);
+            let mut context = Context::new();
+            context.insert("content", &post.rendered_content());
+            context.insert("rel_rootpath", rel_root_path);
+            let output = tera.render("post.html", &context).expect("Failed to build post");
+            fs::write(dst_path, output).expect("Failed to write html file");
 
-            fs::write(dst_path, html).expect("Failed to write html file");
-            println!("{entry:?}")
+            posts.push(post);
         } else {
             copy(src_path, dst_path).expect("Copy file failed");
         }
